entity crd、upd 属性实现自动注入

+++ xljk-moudle/xljk-admin/src/main/java/com/mju/zjj/admin/biz/impl/AuthBiz.java
@@ -35,7 +35,7 @@ public class AuthBiz implements IAuthBiz {
         AdminUser adminUser = adminUserBiz.selectByUsername(username);
         boolean isTrueCaptcha = captchaBiz.verify(uuid, verCode);
         if(isTrueCaptcha&& isTruePassword){
-            return jwtUtil.createToken(adminUser.getId());
+            return jwtUtil.createToken((Integer) adminUser.getId());
         }
         return "";
     }
